+++ src/main/java/com/spotify/docker/client/DefaultDockerClient.java
@@ -56,6 +56,9 @@ import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import com.google.common.io.CharStreams;
 import com.google.common.net.HostAndPort;
+import com.spotify.docker.client.auth.ConfigFileRegistryAuthSupplier;
+import com.spotify.docker.client.auth.NoOpRegistryAuthSupplier;
+import com.spotify.docker.client.auth.RegistryAuthSupplier;
 import com.spotify.docker.client.exceptions.BadParamException;
 import com.spotify.docker.client.exceptions.ConflictException;
 import com.spotify.docker.client.exceptions.ContainerNotFoundException;
@@ -98,7 +101,6 @@ import com.spotify.docker.client.messages.NetworkConnection;
 import com.spotify.docker.client.messages.NetworkCreation;
 import com.spotify.docker.client.messages.ProgressMessage;
 import com.spotify.docker.client.messages.RegistryAuth;
-import com.spotify.docker.client.messages.RegistryAuthSupplier;
 import com.spotify.docker.client.messages.RegistryConfigs;
 import com.spotify.docker.client.messages.RemovedImage;
 import com.spotify.docker.client.messages.ServiceCreateResponse;
@@ -2691,17 +2693,6 @@ public class DefaultDockerClient implements DockerClient, Closeable {
       return this;
     }
 
-    public DefaultDockerClient build() {
-      if (dockerAuth && registryAuthSupplier == null && registryAuth == null) {
-        try {
-          registryAuth(RegistryAuth.fromDockerConfig().build());
-        } catch (IOException e) {
-          log.warn("Unable to use Docker auth info", e);
-        }
-      }
-      return new DefaultDockerClient(this);
-    }
-
     /**
      * Adds additional headers to be sent in all requests to the Docker Remote API.
      */
@@ -2712,6 +2703,23 @@ public class DefaultDockerClient implements DockerClient, Closeable {
 
     public Map<String, Object> headers() {
       return headers;
+    }
+
+    public DefaultDockerClient build() {
+      if (dockerAuth && registryAuthSupplier == null && registryAuth == null) {
+        try {
+          registryAuth(RegistryAuth.fromDockerConfig().build());
+        } catch (IOException e) {
+          log.warn("Unable to use Docker auth info", e);
+        }
+      }
+
+      // read the docker config file for auth info if nothing else was specified
+      if (registryAuthSupplier == null) {
+        registryAuthSupplier(new ConfigFileRegistryAuthSupplier());
+      }
+
+      return new DefaultDockerClient(this);
     }
   }